Split `script` into multiple modules
This should move things around without making any changes to the implementation (other than updating
names). Here is what changed:

- `script::Opcode` is now `::Opcode`
- `Control` `Disabled`, `Operation`, and `PushValue` have moved from `script` to `opcode`
- `LargeValue` and `SmallValue` have moved from `script` to `opcode::push_value`
- `LOCKTIME_THRESHOLD` maved from `script` to `interpreter`
- `script::Script` is now `script::Code` (this is in anticipation of typed
  scripts, which introduces a `struct Script<T> { sig: script::Sig<T>, pub_key:
  script::PubKey }`)
- `script::parse_num` is now `num::parse`
- `script_error::ScriptNumError` is now `num::Error`
- `script::MAX_SCRIPT_ELEMENT_SIZE` is now `opcode::LargeValue::MAX_SIZE`
- `script::serialize_num` is now `num::serialize`
- `script::MAX_SCRIPT_SIZE` is now `script::Code::MAX_SIZE`
- `script::Script::get_op` is now `::Opcode::parse`
- `script::Script::get_lv` is now `opcode::push_value::LargeValue::parse`
