+++ src/interpreter.rs
@@ -9,17 +9,16 @@ use sha2::{Digest, Sha256};
 
 use crate::{
     external::pubkey::PubKey,
-    script::{
-        parse_num, serialize_num, Bad,
-        Control::{self, *},
-        Opcode,
-        Operation::{self, *},
-        ParsedOpcode, PossiblyBad, PushValue, Script, LOCKTIME_THRESHOLD, MAX_SCRIPT_ELEMENT_SIZE,
-        MAX_SCRIPT_SIZE,
-    },
-    script_error::ScriptError,
-    signature,
+    num,
+    opcode::{self, Control::*, Operation::*},
+    script,
+    script_error::*,
+    signature, Opcode,
 };
+
+/// Threshold for lock_time: below this value it is interpreted as block number,
+/// otherwise as UNIX timestamp.
+const LOCKTIME_THRESHOLD: i64 = 500_000_000; // Tue Nov  5 00:53:20 1985 UTC
 
 pub const MAX_OP_COUNT: u8 = 201;
 
@@ -84,7 +83,7 @@ pub trait SignatureChecker {
         &self,
         _script_sig: &signature::Decoded,
         _vch_pub_key: &[u8],
-        _script_code: &Script,
+        _script_code: &script::Code,
     ) -> bool {
         false
     }
@@ -239,7 +238,7 @@ fn is_sig_valid(
     vch_sig: &[u8],
     vch_pub_key: &[u8],
     flags: VerificationFlags,
-    script: &Script<'_>,
+    script: &script::Code,
     checker: &dyn SignatureChecker,
 ) -> Result<bool, ScriptError> {
     // Note how this makes the exact order of pubkey/signature evaluation distinguishable by
@@ -285,8 +284,8 @@ fn binbasic_num<R>(
     op: impl Fn(i64, i64) -> Result<R, ScriptError>,
 ) -> Result<R, ScriptError> {
     binfn(stack, |x1, x2| {
-        let bn2 = parse_num(&x2, require_minimal, None).map_err(ScriptError::ScriptNumError)?;
-        let bn1 = parse_num(&x1, require_minimal, None).map_err(ScriptError::ScriptNumError)?;
+        let bn2 = num::parse(&x2, require_minimal, None).map_err(ScriptError::NumError)?;
+        let bn1 = num::parse(&x1, require_minimal, None).map_err(ScriptError::NumError)?;
         op(bn1, bn2)
     })
 }
@@ -379,7 +378,7 @@ impl State {
 /// trigger some behavior.
 fn eval_step<'a>(
     pc: &'a [u8],
-    script: &Script,
+    script: &script::Code,
     flags: VerificationFlags,
     checker: impl SignatureChecker,
     state: &mut State,
@@ -387,23 +386,25 @@ fn eval_step<'a>(
     //
     // Read instruction
     //
-    Script::get_op(pc).and_then(
-        |ParsedOpcode {
+    Opcode::parse(pc).and_then(
+        |opcode::Parsed {
              opcode,
              remaining_code,
          }| match opcode {
             // See the documentation of `Bad` for an explanation of this logic.
-            PossiblyBad::Bad(bad) => {
-                if Bad::OP_RESERVED != bad {
+            opcode::PossiblyBad::Bad(bad) => {
+                if opcode::Bad::OP_RESERVED != bad {
                     state.increment_op_count()?;
                 }
-                if matches!(bad, Bad::OP_VERIF | Bad::OP_VERNOTIF) || should_exec(&state.vexec) {
+                if matches!(bad, opcode::Bad::OP_VERIF | opcode::Bad::OP_VERNOTIF)
+                    || should_exec(&state.vexec)
+                {
                     Err(ScriptError::BadOpcode(Some(bad)))
                 } else {
                     Ok(remaining_code)
                 }
             }
-            PossiblyBad::Good(opcode) => {
+            opcode::PossiblyBad::Good(opcode) => {
                 eval_opcode(flags, opcode, script, &checker, state).map(|()| remaining_code)
             }
         },
@@ -413,14 +414,14 @@ fn eval_step<'a>(
 fn eval_opcode(
     flags: VerificationFlags,
     opcode: Opcode,
-    script: &Script,
+    script: &script::Code,
     checker: &dyn SignatureChecker,
     state: &mut State,
 ) -> Result<(), ScriptError> {
     (match opcode {
         Opcode::PushValue(pv) => {
             let len = pv.value().len();
-            if len <= MAX_SCRIPT_ELEMENT_SIZE {
+            if len <= opcode::push_value::LargeValue::MAX_SIZE {
                 if should_exec(&state.vexec) {
                     eval_push_value(
                         &pv,
@@ -468,7 +469,7 @@ fn eval_opcode(
 }
 
 fn eval_push_value(
-    pv: &PushValue,
+    pv: &opcode::PushValue,
     require_minimal: bool,
     stack: &mut Stack<Vec<u8>>,
 ) -> Result<(), ScriptError> {
@@ -487,7 +488,7 @@ fn should_exec(vexec: &Stack<bool>) -> bool {
 
 /// <expression> if [statements] [else [statements]] endif
 fn eval_control(
-    op: Control,
+    op: opcode::Control,
     stack: &mut Stack<Vec<u8>>,
     vexec: &mut Stack<bool>,
 ) -> Result<(), ScriptError> {
@@ -527,9 +528,9 @@ fn eval_control(
 }
 
 fn eval_operation(
-    op: Operation,
+    op: opcode::Operation,
     flags: VerificationFlags,
-    script: &Script,
+    script: &script::Code,
     checker: &dyn SignatureChecker,
     stack: &mut Stack<Vec<u8>>,
     altstack: &mut Stack<Vec<u8>>,
@@ -540,15 +541,16 @@ fn eval_operation(
     let unfn_num =
         |stackin: &mut Stack<Vec<u8>>, op: &dyn Fn(i64) -> Vec<u8>| -> Result<(), ScriptError> {
             unop(stackin, |vch| {
-                parse_num(&vch, require_minimal, None)
-                    .map_err(ScriptError::ScriptNumError)
+                num::parse(&vch, require_minimal, None)
+                    .map_err(ScriptError::NumError)
                     .map(op)
             })
         };
 
-    let unop_num = |stack: &mut Stack<Vec<u8>>,
-                    op: &dyn Fn(i64) -> i64|
-     -> Result<(), ScriptError> { unfn_num(stack, &|bn| serialize_num(op(bn))) };
+    let unop_num =
+        |stack: &mut Stack<Vec<u8>>, op: &dyn Fn(i64) -> i64| -> Result<(), ScriptError> {
+            unfn_num(stack, &|bn| num::serialize(op(bn)))
+        };
 
     let binfn_num = |stack: &mut Stack<Vec<u8>>,
                      op: &dyn Fn(i64, i64) -> Vec<u8>|
@@ -558,7 +560,7 @@ fn eval_operation(
 
     let binop_num =
         |stack: &mut Stack<Vec<u8>>, op: &dyn Fn(i64, i64) -> i64| -> Result<(), ScriptError> {
-            binfn_num(stack, &|bn1, bn2| serialize_num(op(bn1, bn2)))
+            binfn_num(stack, &|bn1, bn2| num::serialize(op(bn1, bn2)))
         };
 
     let binrel =
@@ -610,7 +612,7 @@ fn eval_operation(
                 // Thus as a special case we tell `ScriptNum` to accept up
                 // to 5-byte bignums, which are good until 2**39-1, well
                 // beyond the 2**32-1 limit of the `lock_time` field itself.
-                let lock_time = parse_num(stack.rget(0)?, require_minimal, Some(5))?;
+                let lock_time = num::parse(stack.rget(0)?, require_minimal, Some(5))?;
 
                 // In the rare event that the argument may be < 0 due to
                 // some arithmetic being done first, you can always use
@@ -751,7 +753,7 @@ fn eval_operation(
         OP_DEPTH => {
             // -- stacksize
             let bn = i64::try_from(stack.len()).map_err(|err| ScriptError::StackSize(Some(err)))?;
-            stack.push(serialize_num(bn))
+            stack.push(num::serialize(bn))
         }
 
         OP_DROP => {
@@ -795,7 +797,7 @@ fn eval_operation(
             if stack.len() < 2 {
                 return Err(ScriptError::InvalidStackOperation(None));
             }
-            let n = u16::try_from(parse_num(stack.rget(0)?, require_minimal, None)?)
+            let n = u16::try_from(num::parse(stack.rget(0)?, require_minimal, None)?)
                 .map_err(|_| ScriptError::InvalidStackOperation(None))?;
             stack.pop()?;
             if usize::from(n) >= stack.len() {
@@ -842,8 +844,8 @@ fn eval_operation(
                 return Err(ScriptError::InvalidStackOperation(None));
             }
             let bn = i64::try_from(stack.rget(0)?.len())
-                .expect("stack element size <= MAX_SCRIPT_ELEMENT_SIZE");
-            stack.push(serialize_num(bn))
+                .expect("stack element size <= PushValue::MAX_SIZE");
+            stack.push(num::serialize(bn))
         }
 
         //
@@ -897,9 +899,9 @@ fn eval_operation(
             if stack.len() < 3 {
                 return Err(ScriptError::InvalidStackOperation(None));
             }
-            let bn1 = parse_num(stack.rget(2)?, require_minimal, None)?;
-            let bn2 = parse_num(stack.rget(1)?, require_minimal, None)?;
-            let bn3 = parse_num(stack.rget(0)?, require_minimal, None)?;
+            let bn1 = num::parse(stack.rget(2)?, require_minimal, None)?;
+            let bn2 = num::parse(stack.rget(1)?, require_minimal, None)?;
+            let bn3 = num::parse(stack.rget(0)?, require_minimal, None)?;
             let value = bn2 <= bn1 && bn1 < bn3;
             stack.pop()?;
             stack.pop()?;
@@ -966,7 +968,7 @@ fn eval_operation(
             let mut i: u8 = 0;
 
             let mut keys_count =
-                u8::try_from(parse_num(stack.rget(i.into())?, require_minimal, None)?)
+                u8::try_from(num::parse(stack.rget(i.into())?, require_minimal, None)?)
                     .map_err(|err| ScriptError::PubKeyCount(Some(err)))?;
             if keys_count > MAX_PUBKEY_COUNT {
                 return Err(ScriptError::PubKeyCount(None));
@@ -985,7 +987,7 @@ fn eval_operation(
             assert!(i <= 1 + MAX_PUBKEY_COUNT);
 
             let mut sigs_count =
-                u8::try_from(parse_num(stack.rget(i.into())?, require_minimal, None)?)
+                u8::try_from(num::parse(stack.rget(i.into())?, require_minimal, None)?)
                     .map_err(|err| ScriptError::SigCount(Some(err)))?;
             if sigs_count > keys_count {
                 return Err(ScriptError::SigCount(None));
@@ -1058,7 +1060,7 @@ pub trait StepFn {
     fn call<'a>(
         &self,
         pc: &'a [u8],
-        script: &Script,
+        script: &script::Code,
         state: &mut State,
         payload: &mut Self::Payload,
     ) -> Result<&'a [u8], ScriptError>;
@@ -1076,7 +1078,7 @@ impl<C: SignatureChecker + Copy> StepFn for DefaultStepEvaluator<C> {
     fn call<'a>(
         &self,
         pc: &'a [u8],
-        script: &Script,
+        script: &script::Code,
         state: &mut State,
         _payload: &mut (),
     ) -> Result<&'a [u8], ScriptError> {
@@ -1086,7 +1088,7 @@ impl<C: SignatureChecker + Copy> StepFn for DefaultStepEvaluator<C> {
 
 fn eval_script<F>(
     stack: Stack<Vec<u8>>,
-    script: &Script,
+    script: &script::Code,
     payload: &mut F::Payload,
     eval_step: &F,
 ) -> Result<Stack<Vec<u8>>, ScriptError>
@@ -1094,7 +1096,7 @@ where
     F: StepFn,
 {
     // There's a limit on how large scripts can be.
-    if script.0.len() > MAX_SCRIPT_SIZE {
+    if script.0.len() > script::Code::MAX_SIZE {
         return Err(ScriptError::ScriptSize(Some(script.0.len())));
     }
 
@@ -1134,7 +1136,7 @@ impl SignatureChecker for CallbackTransactionSignatureChecker<'_> {
         &self,
         sig: &signature::Decoded,
         vch_pub_key: &[u8],
-        script_code: &Script,
+        script_code: &script::Code,
     ) -> bool {
         let pubkey = PubKey(vch_pub_key);
 
@@ -1178,7 +1180,7 @@ impl SignatureChecker for CallbackTransactionSignatureChecker<'_> {
 /// Additional validation for spend-to-script-hash transactions:
 fn eval_p2sh<F>(
     data_stack: Stack<Vec<u8>>,
-    script_sig: &Script,
+    script_sig: &script::Code,
     payload: &mut F::Payload,
     stepper: &F,
 ) -> Result<Stack<Vec<u8>>, ScriptError>
@@ -1193,7 +1195,7 @@ where
         data_stack
             .split_last()
             .and_then(|(pub_key_2, remaining_stack)| {
-                eval_script(remaining_stack, &Script(pub_key_2), payload, stepper)
+                eval_script(remaining_stack, &script::Code(pub_key_2), payload, stepper)
             })
             .and_then(|p2sh_stack| {
                 if p2sh_stack.last().is_ok_and(cast_to_bool) {
@@ -1207,9 +1209,9 @@ where
     }
 }
 
-pub(crate) fn verify_script<F>(
-    script_sig: &Script,
-    script_pub_key: &Script,
+pub fn verify_script<F>(
+    script_sig: &script::Code,
+    script_pub_key: &script::Code,
     flags: VerificationFlags,
     payload: &mut F::Payload,
     stepper: &F,
